Fixing some bug. Still one bug (tested)

+++ pongLobby/Player.java
@@ -5,7 +5,7 @@ import javax.swing.Timer;
 
 
 public class Player {
-	Player(long id, long port, InetAddress ipAddress, Lobby lobby){
+	Player(long id, int port, InetAddress ipAddress, Lobby lobby){
 	set_id(id);
 	set_port(port);
 	set_ipAddress(ipAddress);
@@ -27,11 +27,11 @@ public class Player {
 		this._id = _id;
 	}
 	
-	public long get_port() {
+	public int get_port() {
 		return _port;
 	}
 
-	public void set_port(long port) {
+	public void set_port(int port) {
 		this._port = port;
 	}
 
@@ -55,7 +55,7 @@ public class Player {
 	//********************************************
 	private long _id;
 	private InetAddress _ipAddress;
-	private long _port;
+	private int _port;
 	private Timer _timer;
 	private Lobby _lobby;
 }
